requirements.txt
gradio
openai
PyMuPDF
gspread
oauth2client
google-auth
google-api-python-client